AdminEndPoints

+++ CarRental.APIs/DTOs/Rental/RentalToReturnDto.cs
@@ -6,7 +6,7 @@ namespace CarRental.APIs.DTOs.Rental
     {
         public DateTime Start_Date { get; set; }
         public DateTime End_Date { get; set; }
-        public int Total_Cost { get; set; }
+        public decimal Total_Cost { get; set; }
         public string Pick_Location { get; set; }
         public string Ret_Location { get; set; }
         public RentalStatus Status { get; set; }
